feat: divide JS code to several modules
